+++ src/control/AudioController.cpp
@@ -55,7 +55,7 @@ auto AudioController::isRecording() -> bool { return this->audioRecorder->isReco
 
 auto AudioController::isPlaying() -> bool { return this->audioPlayer->isPlaying(); }
 
-auto AudioController::startPlayback(const string& filename, unsigned int timestamp) -> bool {
+auto AudioController::startPlayback(const string& filename, uint64_t timestamp) -> bool {
     this->audioPlayer->stop();
     bool status = this->audioPlayer->start(filename, timestamp);
     if (status) {
