+++ src/control/AudioController.h
@@ -35,7 +35,7 @@ public:
     bool isRecording();
 
     bool isPlaying();
-    bool startPlayback(const string& filename, unsigned int timestamp);
+    bool startPlayback(const string& filename, uint64_t timestamp);
     void pausePlayback();
     void continuePlayback();
     void stopPlayback();
