+++ src/control/SearchControl.h
@@ -19,7 +19,7 @@ public:
     SearchControl(const PageRef& page, XojPdfPageSPtr pdf);
     virtual ~SearchControl();
 
-    bool search(string text, int* occures, double* top);
+    bool search(string text, int64_t* occures, double* top);
     void paint(cairo_t* cr, GdkRectangle* rect, double zoom, const GdkRGBA& color);
 
 private:
